Update patch set 1
Patch Set 1:

> > We need to use known STABLE build of vpp not the latest one
 > > (possibly impacted by some degrades) for test case development.
 > > Otherwise it would be double work to check issues - test case
 > > problem or vpp build problem?
 > 
 > How you know which one is stable? Code changes in vpp repo will be
 > checked by tests before commit to master. Rewrite version to
 > download it is not right.

Well in fact this has two answers. For CSIT verification we indeed need to have version of VPP that is "static" in a sense that it is not changing in between tests. That way we have only one variable in validation run, which is test case changes.

On the other hand, for VPP verify changes, we need to do the same for CSIT version. We are going to have master and golden branch, and the golden branch is going to be updated in periodic manner so that changes in VPP are verified against known working state of CSIT tests.

Patch-set: 1