Update patch set 2
Patch Set 2:

(5 comments)

Zdenko, thanks a lot for piking this up! You have some general comments after a quick non thorough review. An important point to clarify: are these tests for lisp+ipsec? If yes, then we should have that stand out in the name of the test suite. If not, what is the difference with respect to the tests in lisp_dataplen_untagged.robot?

Patch-set: 2
Label: Code-Review=0