Update patch set 5
Patch Set 5:

A large portion of the code that was in bootstrap.sh has been replicated into bootstrap-centos.sh and bootstrap-ubuntu.sh.  I think it would be better if this were refactored to reduce the replication.   Especially for things like the list of server addresses, and the public key, etc.

I understand that time is of the essence in getting the Centos support completed, so I'm ok if the refactoring is postponed and a JIRA ticket opened to address the replication issues separately.

Patch-set: 5
Label: Code-Review=0