Update patch set 5
Patch Set 5: Code-Review-1

I don't know the cause of the failure but I am not going to give this a +1. I tested at least a few LISP and vxlan tests manually on server t1 without any failures but I think we should start upgrading the two production servers to 1604 and latest VIRL/Mitaka and see if these failures disappear.

Patch-set: 5