Update patch set 3
Patch Set 3: Code-Review+1

Peter, it looks good to me and seems to be a straightforward update. I will build it and try to run  it in a stand-alone VM today.

You may want to put DO NOT MERGE in the comment instead of giving it a -2 yourself. -2 makes it look like you found something wrong with it.

Patch-set: 3
Reviewer: Gerrit User 32 <32@6d2eb258-4fe2-443e-8a38-ca81da23d4c2>
Label: Code-Review=+1