Update patch set 1
Patch Set 1:

I assume that this failure is due to the vpp startup problem on Centos which is due to the selinux requirement is not compatible with Centos 7.3.
I think I should abandon this patch and roll it into the general csit Centos image update which has a fix for the above.

Patch-set: 1