Update patch set 10
Patch Set 10: Code-Review-1

> none of 9000B passed

I see the status as PASS, but that is probably just because we do not have any checks that would fail the test on rx=0.

> Can you find out why?

I was unable to find out why. I will send an e-mail to vpp-dev and remove the 9000B test cases.

By the way, here is a bash one-liner to print which test cases had rx=0:
  $ cat consoleText | tr '\n' '_' | sed 's/-_/`/g' | sed 's/=_/`/g' | tr '`' '\n' | fgrep ', rx 0' | cut -d ' ' -f 1 > failed.txt

Patch-set: 10
Label: Code-Review=-1