Update patch set 10
Patch Set 10:

The original idea was to give to developers using MRR to evaluate results themselves. It means that test will fail only in case of critical issue. But it is not a problem to implement the check for 0.
And I think make sense to put there condition.

Patch-set: 10
Label: Code-Review=0