DO NOT MERGE: Put Centos 7.4-1711 1.9 image into production
Testing VIRL3 server by way of patch to debug verify problem on
gerrit 11318.

The patch to create this image is gerrit patch 11245

CSIT-566
CSIT-688
CSIT-681
CSIT-682
CSIT-688
CSIT-683

7.4-1711 1.9 images have been installed on VIRL1, VIRL2 and VIRL3

The 7.4-1711_1.9 image has been tested locally on VIRL3 with a local patch
modifying bootstrap.sh. A pass of verify should force a check on VIRL2 and
VIRL1 to verify that the image works OK.

A merge of this patch will put the new image into production so review this
patch plus 11245 above carefully before merging.

Change-Id: I74a0b23d75120fdc3934bfe3c6e96f8fb10784db
Signed-off-by: Thomas F Herbert <therbert@redhat.com>