Update patch set 2
Patch Set 2:

> > PIP is installed separately in disk-image builder
 > 
 > So its version is not visible here. We could add comments into
 > requirements.txt pointing readers to disk-image builder
 > configuration, but I would prefer "pip install pip==8.1.1" here to
 > make the requirements self-contained.
 > 

Ok, i do not see a problem to add to list.

 > > there is no way "some" package will be upgraded
 > 
 > I looked at a recent console output [0] and I have some answers.
 > Apparently the jobs do not use this requirements file, but another
 > one, I think resources/tools/disk-image-builder/ubuntu/lists/ubuntu-16.04.1_2017-10-21_2.0/pip-requirements.txt
 > Good news, that file lists the complete dependency forrest
 > (including docopt and python-dateutil), so no hidden upgrades can
 > happen.
 > I see the following packages installed at virtualenv creation, not
 > sure how to verify their versions: setuptools, pkg_resources, pip,
 > wheel.
 > Also, --system-site-packages is used, so some packages are not
 > downloaded. In case of scapy, the system installed version is
 > 2.2.0, but I hope the downloaded version 2.3.1 is installed
 > correctly and takes precedence.
 > 
 > I would like the whole CSIT to use just one virtual environment
 > with one requirements.txt, where everything is listed explicitly,
 > with comments eplaining which part (run_cpta.sh, run_doc.sh, ...)
 > needs which package.
 > Currently we have 23 requirements.txt files in the git repo.
 > 
 > Current state is not good, but at least it is tested. -1 from me
 > for now.
 > 
 > [0] https://jenkins.fd.io/view/csit/job/csit-vpp-perf-check-master/197/console

Most probably we need another TOI on how this works. 
Shortly. Few of them are for PAL (separate part of code). The rest is part of disk-image-builder (DIB). DIB is statically linking "backup" of pip as part of image creation (for auditing purposes), this is done during image compilation. That is why verify job is now failing. It is not installing "this" pip, but using its previous version on VIRL/PhyTB. Only Jenkins slaves are using this new list. To be able to use new pip list i need to rebuild VIRL image and update PhyTB. New version of VIRL image will be produced once agreed on this list (avoiding rebuilding many times). Once new image - i will re-run this gerrit patch for verify.

Patch-set: 2