Update patch set 1
Patch Set 1:

> Who is maintaining the list here?

Csit committers.

> Why fd.io/csit commiters and csit contributors are not on the list?

Contributors should not be there (should be fixed by promoting me to committer status).

Committers can be there. The complication is that LF identity is not in any way tied with PyPI account. I do not think every csit committer should be required to create PyPI account just to be on the list. Of course, the more owners the safer the code maintenance is.

Patch-set: 1