Compatibility fixes with Ubuntu 18.04
Currently we are using ip:port socket to for VPP to listen for incoming
connections. There is only one place in code this socket is used for
detection if VPP is responding after restart. This patch is supposed to
change ip:port to sock:file which is default preffered way in VPP. This
way we can start using vppctl. netcat on the other side with ip:port
remains hanged in console for infinite time not returning back.

There is also need to check whether uio_module is loaded and eventually
load it. As uio_pci_generic is not working there, we will fallback to
main igb_uio (uio_pci_generic being only lightweight subset of igb_uio).

This patch also add additional verbose output to console during setup to
better understand if failure happens during copy/extract/setup.

Change-Id: I0b8a3b76dce3316496a3ad928f58718b6e73057d
Signed-off-by: Peter Mikus <pmikus@cisco.com>