Update patch set 48
Patch Set 48: Code-Review-1

(6 comments)

Generally look ok, but i think this is not PAL ready as output is changed and I am not sure removing TX from output is right way. As we do not know what was the Transmit rate (sure one can compute from the NIC but e.g. XL710 is not that straight forward). Lets discuss also displaying as it is not clear what is it displaying:
e.g.: Maximum Receive Rate Results [8015506] ??? Seconds, packets percentage of line rate? units are missing it is not clear if it is agreggated and if it is PPS/FRAMES total ... without reading code.

Please make sure you will run at least few tests with current NDR/PDR as they are using perf_trial as well. We cannot afford disabling them yet.

Patch-set: 48