Remove THIS tag and use unix endlines
The THIS tag has sneaked in 10Ge2P1X520-Eth-L2Xcbase-Eth-1Memif-1Dcr-Mrr
but usually it is only to be used in draft testing.

Not sure about the way non-unix endlines crept in.
Is not git supposed to perform endline conversions on text files?
Or are .robot, .rst, .yaml and .vat not recognized as text?

Change-Id: Ia78cf942c05c3029c340a8639782786f37e57469
Signed-off-by: Vratko Polak <vrpolak@cisco.com>