CSIT-1186: Add srv6 MLR suites
Also, there is a script to re-generate testcases of selected suites.
For that, resources/libraries/python/autogen is created,
with Testcase for customizable tescase template
and DefaultTestcase for a simple template.

The script looks at files matching a glob expression,
reuses a "prolog", but overwites test cases with the default template
and hardcoded set of testcase arguments.

Change-Id: I394ad61f262b3078e5ca719034b15ada48987ee0
Signed-off-by: Vratko Polak <vrpolak@cisco.com>