Update patch set 23
Patch Set 23:

> Looking at https://jenkins.fd.io/job/csit-vpp-perf-verify-master-3n-hsw/10/
 > the new overhead computations looks ok.

The basic key point there is that once you reach MTU you are allocating chain buffers in VPP. In CSIT we need to be sure that A) we are enabling no-muliti-seg strictly only when pkt size < 1522B, B) we never oversubscribe NIC so max rate on the links must be calculated on links that contains encapsulation. as all encapsulation are done between DUTs (that is why CSIT has 3-node) we need to calculate with this overhead to have PPS shaped properly. I will  look on logs, if there is no bug on CSIT side and test is properly set, this could indicate the problem that e.g. SRv6 vpp graph nodes are broken for Jumbos. As basically all IMIX/1512/9000B are here considered as jumbo (due to pkt overhead) and we are sure we did not miss anything we can open VPP ticket.

Patch-set: 23