Update patch set 71
Patch Set 71:

> I am trying to install libnuma-dev on DUT's,
 > but it is failing
 > 
 > https://jenkins.fd.io/job/csit-dmm-functional-master-ubuntu1604-virl/91/console
 > 
 > E: Unable to locate package libnuma-dev
 > 
 > How can I fix this.
 > can someone help me in this.

It was mentioned earlier. DUTs do not have internet connectivity, if it is not pre-cached in apt it is not there. There are only 2 options. 1) Compile everything on Jenkins and tranfer final package to all DUTs (this is the same way as VPP is being tested - we are not compiling VPP in duts) 
2) New VIRL image needs to be spinned with all requirements - this could take some time as it is not now aoutomated on Jenkins level.

Patch-set: 71
Label: Code-Review=0