Update patch set 26
Patch Set 26:

> (1 comment)

Same I can say can be done with opposite approach. If you expect some piece of code to use undef variables or exit in any case make that explicit and set safer -e or -u. Specifically boostraps are not those we want to quit imediately from various reasons (see below).

So question is more like what coding style we want to use further on. Explicit or implicit. With Explicit it will fail but coder needs to make sure that it will not fail when we do not need it (that is like 70% of boostrap code needs to be excluded - kubernetes/reservation/archiving/downloading/virtualenv maybe others). Implicit would be reverse logic. So fail the code when needed. This or that requires same effort i think. 

Regarding ${variable:-default} i think this is not the case. Look at line 369 you want to copy aritifacts to Archive just only if variable is set (no defaults) so another exception from unset otherwise boostrap will not work on localhost.

Patch-set: 26
Label: Code-Review=0