Update patch set 78
Patch Set 78:

> >>>Still waiting for gitignore.
 > the files we are editing here are not from csit repo
 > the scripts that are being edited are from dmm repo(cloned inside
 > csit)
 > https://gerrit.fd.io/r/#/c/14375/77/tests/dmm/dmm_scripts/run_dmm.sh
 > : 23
 > I thought u meant about dmm .gitignore

Looking on your discussion it is not clear to me why #23 line cannot be just replaced with parameter. If the purpose of run_dmm.sh is to be reused in other project/packages/scripts then mae the "hugepages" as parameter and you do not need to do obscure sed-ing?

Patch-set: 78
Label: Code-Review=0