Update patch set 2
Patch Set 2:

> Also, I am not sure how much our installation procedure relies on
 > vpp-ext-deps (as opposed to official packages from www.dpdg.org or
 > similar), so we should verify on multiple NICs.

It does not rely anyhow. I had an argumentation fight with Damjan explaining him we are integration testing and we just revealed that vpp1810 release is not complete, but unfortunately he wants to remove this reportedly "useless" package. And i do not mind to maintain this package anymore as it just making headache for any CSIT dev for very long time :)

Patch-set: 2