Update patch set 2
Patch Set 2: Code-Review-1

Per my comment on slack:

I like l2xc to be tested but if the preference is to have as many test as possible than i am fine. But natural question then is: if VHOST, why not also Memif. We are allowing to vpp-dev to use custom query. so they can add more test if needed. But what is baseline here is i think pure l2xc, not vhost as vhost is feature set.
But i am not strict about that, it just make more sense to me.

Similarly if ipsec developer need ipsec test in baseline set that what are you going to change that? I this because of Damjan's patch?

Patch-set: 2
Reviewer: Gerrit User 76 <76@6d2eb258-4fe2-443e-8a38-ca81da23d4c2>
Label: Code-Review=-1