Update patch set 1
Patch Set 1:

> !9000b

Is it known that our DPDK tests do not work with jumbo frames? Can we fix the tests, or is it the DPDK app limitation?
If it is a known limitation, we should update the suite generator to not emit 9000b test cases.

Patch-set: 1
Reviewer: Gerrit User 1263 <1263@6d2eb258-4fe2-443e-8a38-ca81da23d4c2>
Label: Verified=0