Update patch set 1
Patch Set 1:

> > tell the difference
 > 
 > I am in favor of stricter checks. If CSIT does not really need to
 > be updated, bumping API_SIGNATURE is very quick.

This is the model that Ligato team was following and after months they had to go to less strict because from long term point of view it is very resource eager (personal) to maintain it. I can teel it will cost 0.5HC allocated to just do API. 

We can discuss on public call.

I do +1 you idea, because it is right thing, but Maciek will not be happy (in fact any PTL).

Patch-set: 1