Update patch set 1
Patch Set 1:

Not changing and being backwards compatible is clearly a grand goal.
The downside of it is that it might hinder evolution, it might incur a big tax on making changes.
Right now the balance is skewed against the users of VPP. You might argue that the pain is constant, and it's purely a matter of deciding who is going to suffer.
There might not be a great solution here, but could we try to find one? Accept that APIs will change, and then try to come up with accommodating schemes? If that's auto-generated code. Self-describing data structures or what not I don't know. Clearly we cannot isolate ourselves from all changes, but perhaps from most?

Patch-set: 1
Label: Verified=0