Update patch set 62
Patch Set 62:

> just better select NICs

The thing is, the previous expressions were relying on some suite+nic combinations missing, so it is not easy to tell which surprising tests would manually created expressions allow.

In other words, simplification of tag expressions will take a long time, but I would prefer for that effort to be done after this is merged (so I do not have to keep rebasing for stuff like 18157).

Patch-set: 62