Update patch set 81
Patch Set 81:

(5 comments)

Hard to review the logic as the patch is huge, hope the tear downs refactor works as expected similar to other parts like renaming of variables (which are easy to check by fgrep).
The other dangerous area are NIC limits where hopefully this simplification will cover all caveats that comes from variety of combination in LF and outside.

Patch-set: 81