Update patch set 19
Patch Set 19:

(1 comment)

The nomenclature of the signature directory names (gerrit, origin) are not very intuitive.  Both directories contain API signatures which are supported by the CSIT version in the branch in which they exist.

In theory there could be more than two signatures that are supported by a given CSIT version (ie. in the case where multiple api files were changed that are not used by CSIT, or which contain backwards compatible modifications).

I'd prefer to make it more generic, for example:
vpp-api-client-signature
vpp-api-client-signature.1
. . .
vpp-api-client-signature.n

If any of the supported signatures match the one in the patch, then the testing can proceed.  If none of them pass, then Verified -1 and initiate the flag day dance.

Once a CSIT branch is changed to be incompatible with a given signature, then the associated signature directory shall be removed.

In practice, it makes sense to limit the number of signatures supported to remove the need to verify all of the signatures for CSIT patches.  In the short term, this is likely to be a manual process, but once CSIT is converted to using PAPI everywhere, then it should be able to be automated by having the PAPI executor generate a log of all of the PAPI calls executed (including all attribute values used!).

Patch-set: 19