Update patch set 5
Patch Set 5:

Long comment. Mostly written before I realized the underlying question is which party (VPP or CSIT) should be more motivated to prevent breakages.

> Was the communication broken and how having a block would have changed it ?

The communication is still broken. I only look at merged VPP changes, so I do not see comments on unmerged VPP changes, unless I am added as a reviewer (or the Change is mentioned on vpp-dev).
The block would take a hostage, in order to motivate VPP developers to start (and continue) notifying CSIT developers.

> Ole said "Let's see what CSIT says"

In the particular case of 18352, the device job was already showing failure in all its tests.

> A voting job does not start a negotiation, it just takes a hostage

I agree. Apparently I am not seeing the feature of hold-down period that is better in starting the negotiation.

Is it the automated e-mail? That is also present in the "flag day" proposal.

Or do you think CSIT developers fearing the hold-down period expires before CSIT is ready is a better motivator than VPP developers fearing their Change will never get +1 from CSIT job? That could actually be true.

And it would also explain why hold-down period would be useful for other consumers (not just for the privileged ones).

>> I do not think hold-down period would force enough focus.

> Why would it not?

Because VPP developers can wait it out without engaging in communication.

> so my CI could continue

Your CI does not have the leverage comparable to CSIT trending.

By the way, the changes affecting "make test" only are not covered by the current "flag day" proposal, as those tests are not packaged for CSIT to see.

Patch-set: 5
Label: Code-Review=0