Update patch set 294
Patch Set 294:

> I see (not a VPP crash):
> 
> error: [Errno 111] Connection refused

Yeah. VPP refused connection attempts for some time, but then decided to resume responding. I could increase the number of tries, but that is just a hiding the issue of VPP communication not being stable. I prefer to keep such occasional failures, so that we can go back and investigate them properly later.

Patch-set: 294