Update patch set 9
Patch Set 9: Code-Review+2

I created [0] to trigger the conditional parts of teardown. It shows [1] that our efforts at reducing the logging was successful, with unintended consequence of "show" keywords not actually showing the results.
I believe this would have to be addressed for multiple "show" keywords, so this is good enough.

[0] https://gerrit.fd.io/r/20750
[1] https://logs.fd.io/production/vex-yul-rot-jenkins-1/csit-vpp-perf-verify-master-3n-hsw/489/archives/log.html.gz#s1-s1-s1-s1-s1-t1-k3-k5-k2-k1-k1-k1-k1-k2-k1-k1

Patch-set: 9
Label: Code-Review=+2