Update patch set 10
Patch Set 10:

(1 comment)

resources/libraries/python/Policer.py:
Line 26:
From the dictionary arguments, i see that rate_type is "pps" which is a string. I also understand that the CLI command's requisite is also a string (pps).

With that in picture, could you please elaborate as to why the lowercase strings (Line No - 26) should be converted to ints or used as tuples.

Patch-set: 10