Update patch set 10
Patch Set 10: Code-Review+1

Hi Team,

Thanks for your comments.

I have converted the lowercase strings to integers (eg. kbps to 0, pps to 1) and so on with the help of xlate.h

Whatever parameter is used as u8 in policer.api,i have set them to be an integer in Policer.py

Please let me know for clarifications.

Regards,
Prem.

Patch-set: 10
Label: Code-Review=+1