Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> > We can move API check after second restart where we own the config.
> 
> No, we need to make sure VPP is up, so restart will replace it indeed.
> Remember that change which made VPP ignore SIGINT when workers are not initialized yet?

I do remember, but we have a race condition, if we want to override sock we cannot because default statup.conf has fixed location. Supporting default and overridden value is fine but lots of code. PAPISocket would need to support default (which can change) and then new which is known by CSIT. This is equal to API change in fact. I do hope they will not change soon. But for these reasons we need vpp_device to be voting, so we can prevent. It is not enough to just check CRC if CSIT cannot use it (e.g. startup crc check). 

This why i was thinking that whole concept of checking CRC is not enough. It would be better to have every API function in CSIT to register itself (directly in function, not the list as of today) and then run a simple one TC which will check registrations and if they match installed version. This way it would really simulate client-to-vpp + it is fully automatic in the way that registration does not need a separate list.

Job would finish in less then 5minutes as it would just check registrations + vpp startup.

The only problem is implementation time as of now it is too late.

Patch-set: 5