Update patch set 1
Patch Set 1: Code-Review-1

It fails [0] and pylint sees [1] an undefined variable.

If the code was simpler, I would argue this is a pure refactor to de-duplicate codebase (as pylint complains about it).
But I do not like the _needs_ methods, so perhaps the bigger refactor (to use bash for running any software installation) is a better idea.
There are easier pylint violations to fix in the meantime.

[0] https://logs.fd.io/production/vex-yul-rot-jenkins-1/csit-vpp-device-master-ubuntu1804-1n-skx/3155/archives/log.html.gz#s1-s1-s1-k1-k2
[1] https://jenkins.fd.io/job/csit-verify-tox-master/1915/violations/file/resources/libraries/python/SetupFramework.py/

Patch-set: 1
Label: Code-Review=-1