Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> It fails [0] and pylint sees [1] an undefined variable.
> 
> If the code was simpler, I would argue this is a pure refactor to de-duplicate codebase (as pylint complains about it).
> But I do not like the _needs_ methods, so perhaps the bigger refactor (to use bash for running any software installation) is a better idea.

WHAT I DISLIKE:
This is why consulting before is better idea then do code and then redoing/scratching_head. Plus these tests are dead for more than half a year with no maintenance for more than year. There is currently no guarantee of anything than making this from scratch. This is what I commented earlier it must be done with lot of planning ahead. Testing TLDK with 16.xx is like for what --there is no audience?!!! This is general failure of community to make these tests up to date. NSH was adopted as plugin this must be reworked totally. SFC is still supported but one should do inspection first what are the plans of testing of up to date branches.
Best would be to keep setup framework not specific for parties as they are failing to update and we need to move on. So making SetupFramework fixed is the better idea -- no changes for sub-projects.

WHAT I LIKE:
Unification and breaking into class of main SetupFramework looks more mature code then before and this can be separately adopted (once working of course).

GENERALLY:
All three sub-project integrations are done in very specific way making them very hard to maintain for csit-core small team. Better way is to keep SetupFramework more monolithic and doing CI/CD integrations elsewhere where it will be easier for other parties to maintain or remove if not needed. Also there is problem of lack of tests. There are some sample TCs clearly not covering a fraction of functionality with no clear trajectory of adding more or removing not needed.

Patch-set: 1