Update patch set 1
Patch Set 1:

I still owe the feedback on this but I will try to think about AI as soon as possible. Change is not harmful but fast loop means we do not even need to delete from list as it can be the same situation. Need to think deeper.

Patch-set: 1