Update patch set 6
Patch Set 6:

From PAL point of view it looks good.

But the console output and also the test message in the log becomes unreadable. I suggest to remove latency from test message (and console output) and replace it with a keyword which will provide all output data as a json string.

Patch-set: 6
Label: Nomad=0