Reconf tests: Fix async measurements
TRex does not zero the server counters.
It copies the values to use as reference,
and subtracts them when asked for results.

But the reference is stored in the client (not the server).
And CSIT uses different scripts to start and stop async traffic,
which means different clients.

This patch introduces a workaround.
Async start will return xstats objects to use as reference,
and async stop will use the objects to compute the correct results.
The xstats objects are stored in TrafficGenerator instance.
Sync measurement does not export the counters, to shorten logs.

Other improvements:
+ Make stop_traffic_on_tg return measurement results directly.
+ Rename --async to --async_start as "async" is reserved in Python 3.7
+ Minor pylint, docstring and typo fixes.

Change-Id: I5fc56a0763afb7d62cfa7c0651f96b6867de3e15
Signed-off-by: Vratko Polak <vrpolak@cisco.com>
(cherry picked from commit bab0b570345ceb6ffeaec9e47a50e62d7303387e)