Update patch set 3
Patch Set 3: Code-Review-1

> csit-verify-tox-master/2532/ : SUCCESS

I see a log file created [1], but empty.
And I see the following in the console [2] output:
  xargs: shellcheck: No such file or directory

I believe that means the Ubuntu package providing "shellcheck" command is not installed.
As the executor is a temporary container, I believe it is ok to add "sudo apt-get install shellcheck" while debugging, but the final version should reply on the executor having everything needed.

[1] https://logs.fd.io/production/vex-yul-rot-jenkins-1/csit-verify-tox-master/2532/archives/shellcheck.xml.log
[2] https://logs.fd.io/production/vex-yul-rot-jenkins-1/csit-verify-tox-master/2532/console.log.gz

Patch-set: 3
Label: Code-Review=-1