Hotfix: Skip socket check in global perf setup
On Taishan, socket check leads to quite reliable fail,
so all tests are skipped.
Inserting sleeps does not fix the behavior reliably.

A better way would be to add a test for the Taishan failure,
but we have missed enough trending data already.
Also, the current bisect script does not work
if CSIT code leads to test failures.

+ Do not retry on socket connect failure,
  there are no intermittent failures it would help with.

Change-Id: I87a5f5230914bf8636d72fff362be9564d1ee4df
Signed-off-by: Vratko Polak <vrpolak@cisco.com>