Update patch set 1
Patch Set 1:

the issue of this patch is that it is not solving the baseline problem. See the tests that are failing for referred VPP patch and explain how IP6 plugin is affecting ipv4 tests.
Also making "ping_plugin.so" should not be required when configuring ARP statically (because i cannot run VPP without activating it - as it fails for static environment as well).

Patch-set: 1