Update patch set 2
Patch Set 2:

> I prefer the previous way to have exact list what we are running

We have lists in ci-management, which decide which jobs are running on which platforms.

What I do not like is having multiple lists, when forgotten update results in failure only on some infrequently used jobs.

> avoid sed everywhere as it tends to be platform dependent

I think the subsequent "sed -oE" is a bigger risk.
Perhaps there is a better way, but I am not seeing it yet.

Patch-set: 2