Improve diagnostic output in reservation scipt
Shorten the output for usual situations
(reservation dir missing or containing a file).
Print exact command and output (including time data)
if reservation dir is empty,
or if "ls" output is not parseable for the file name.

When the file name has a particular form,
output a URL so people can click to see the owner run.

Do not run "ls" after reservation failed,
as all the info should be visible already.
Race conditions can still happen,
but the info is important only on jobs
which fail to reserve for a significant time.

Change-Id: I7bbbdb4f3ef433f18ef236697e043bc04ad47b4a
Signed-off-by: Vratko Polak <vrpolak@cisco.com>