Update patch set 5
Patch Set 5:

I am one of the consumers of reservation output.
Usually I want to know how quickly will my triggered job will get a testbed.
If the run gets it, I do not care about output. But the run is stuck waiting, I want to know why, so that I can raise an alarm when something does not work as expected.

> > 1. See timestamps when job tag is not written.

> 1. No timestamp, it is pointless in day to day operation, nobody is going to investigate who did reservation earlier or what was the linux epoch time. I do not care, i want simple one file only no garbage. And if I will reserve TB for any purpose you will not be wiser at all.

When job tag is not written, it means there is no "simple one file" present. If it is recent enough, I do not care much (hoping it is just a manual reservation gone wrong).
If the situation persist for hours and days, I raise more specific alarms if I know the timestamp.
Also without timestamp, I know how long "situation persists" only with respect to my trigger, not with respect to the date reservation dir has been updated/created.
(Also, the timestamp is human readable, while "epoch time" is usually just an opaque number.)

> > 3. Add URLs so I can save time (without breaking s5ci).

> 3. No URLs, again you will not be wiser then today ...
> Also reservation can be manual, so i can redirect you anywhere.

I will be wiser faster.
Some run times are as long as planed, some can be long due to specific triggers. Some are long because they bugged out. I need to navigate to the run page to see the details.
The code in this Change attempts to parse the file name, and prints nothing if the structure is not right.
This point is about speed (how quickly can I check the run). Your ability to redirect manually remains unchanged.

> > 4. See complete command and output if unexpected situation happens.

> 4. Again what will you see apart that reservation cannot be done due to dir exits?? --nothing

This is about "ls" line not containing the space before the file name.
Should not happen, so if it happens I want to see the details (we could have some bug in the reservation script, or somebody is really creative with file names).

> You are just adding overhead to code with no added value, spending time on it and you will not be able to see if TB is reserved for purpose if something goes bad.

Values added:
1. More debug information for some testbed failure scenarios.
3. Less time wasted when investigating potentially stuck runs.
4. Debug information for script not behaving as expected.

> and I am not going to cripple code for S5ci that was tested.

Ok, inability to test other usages (without wasting too much time) is an argument.

Hopefully in the future, contributors will be able to test how do their improvements work in all related usages (otherwise we are stuck in the current implementation).

Patch-set: 5