Update patch set 5
Patch Set 5:

> we could have some bug in the reservation script, or somebody is really creative with file names

Reservation script works more that 3,5years and it is reliable in our conditions. It fails to unreserve only in case of failure of jobs, which i know but this is prevention for major damages to systems (io deadlocks).
So no point to improve something that works and something that is reliable as anyway it requires manual  intervention to fix, when the fix is not really fix just verbosity view that is limitless in compare to possible root causes.
100% of time when machine is locked manual intervention must follow and the fix usually cannot be automated anyway (like NW outage, HW outage, Jenkins outage).

Patch-set: 5