Update patch set 11
Patch Set 11:

> script reveal the issue with 3n-skx (why is it reserved)

Thanks to knowing the date, I was able to find the console log [0].
DUT1 has become unreachable, DUT2 reported "Failed to import docker or docker-py".
As in a previous run with similar ansible failure (both DUTs saying [1] `lxc` module is not importable), the testbed remained reserved, hoping for human intervention.
I guess the intervention happened, but at the end used wrong "rm" command, deleting the file but not the directory.

I have prepared a Change [2] to switch the behavior on ansible failure (allowing the unreservation to finish), if that is what we want.

[0] https://logs.fd.io/production/vex-yul-rot-jenkins-1/csit-vpp-perf-mrr-daily-master-3n-skx/696/console-timestamp.log.gz
[1] https://logs.fd.io/production/vex-yul-rot-jenkins-1/csit-vpp-perf-mrr-daily-master-3n-skx/695/console-timestamp.log.gz
[2] https://gerrit.fd.io/r/c/csit/+/23559

Patch-set: 11