Update patch set 4
Patch Set 4:

(14 comments)

While my code review comments reflect the current usage of u"" string literals, I have a doubt that this is required.  Per [0] and [1], the u"" string literal prefix is not required in Python3 and actually was removed in Python 3.0, then restored in Python 3.3 in order to make the conversion of unicode aware applications less work.

In particular, the Proposal section of PEP-414 [1] specifically states that there is no difference between "foo" and u"foo".

I read a useful explanation [2] of how Python3 uses Unicode code-points internally introduced in Python 3.3 and why PEP-393 [3] is a good idea.

Thanks,
-daw-

[0] https://docs.python.org/3.1/whatsnew/3.0.html#text-vs-data-instead-of-unicode-vs-8-bit
    "You can no longer use u"..." literals for Unicode text."
[1] https://www.python.org/dev/peps/pep-0414/
[2] https://www.b-list.org/weblog/2017/sep/05/how-python-does-unicode/
[3] https://www.python.org/dev/peps/pep-0393/

Patch-set: 4