Update patch set 1
Patch Set 1:

This has one simple problem. Limit 37Mpps is to high when using 2p1nic although fit the 1p1nic. You need to test this limit on 2p1nic and see how long will be T-Rex sending traffic and share the results here.
The problem with trex was that it tries to send packets and wait for them so instead of duration 10s you will end up in 30s. PLease test and share the results here.

Patch-set: 1