Update patch set 1
Patch Set 1:

We know trex has some (duration and load independent) overhead, for example on x710 measurement of 1.0 second took [0] 1.3 seconds to finish. But on xl710 (with this patch), it took [1] 2.2 seconds.
The number of packets received was 27921907 vs 49152838.

I will manually trigger a run with longer trial duration to diminish the effect of the overhead.

[0] https://logs.fd.io/production/vex-yul-rot-jenkins-1/csit-vpp-perf-verify-master-2n-skx/959/archives/log.html.gz#s1-s1-s1-s1-s1-t1-k2-k9-k1-k6-k1-k1
[1] https://logs.fd.io/production/vex-yul-rot-jenkins-1/csit-vpp-perf-verify-master-2n-skx/959/archives/log.html.gz#s1-s1-s1-s1-s2-t1-k2-k9-k1-k6-k1-k1

Patch-set: 1