Update patch set 11
Patch Set 11:

> when the fix is not really fix just verbosity view

This Change (as opposed to 23559) is not a fix to some error in the current process. It is just  an improvement on the diagnostic output, so it gives more info when something is wrong, and less info (plus URL to speed me up) if everything is within expectations.

>>>> unless your script reveal the issue with 3n-skx (why is it reserved)
>>>
>>> you already have timestamp in actual code
>>
>> No timestamp to be seen.
>
> Exactly what we need.
>
>>> that job did not cause the failure
>>
>> a guess about human intervention
>
> Again that was not the job.

Is this some kind of a riddle? If yes, I am giving up on guessing.

I am still curious though. What was the cause? Why is the testbed still not fixed?

Slightly off-topic:
> Reservation script works more that 3,5years and it is reliable in our conditions

Ignoring stuff that got reverted later, the feature of the tag file being stored in the reservation directory (and printed before any reservation attempt) is currently just under 5 months [0] old.

Unless you are talking about the critical section based on the atomic mkdir/rm, which is of course fine (and not touched by this Change).

(from 23559)
> more noticeable is higher Q in jenkins then failed jobs that no one sees

Daily job [1] does not look so bad, only one run is queued currently, but that is just because the job has "concurrent: false". Run #703 took so long due to waiting for the weekly run.

But I am hurting, as my bisect script (say 8 reservations, as it unreserves while compiling) takes days [2] to finish.

[0] https://gerrit.fd.io/r/c/csit/+/20209
[1] https://jenkins.fd.io/view/csit/job/csit-vpp-perf-mrr-daily-master-3n-skx/buildTimeTrend
[2] https://jenkins.fd.io/view/vpp/job/vpp-csit-verify-perf-master-3n-skx/buildTimeTrend

Patch-set: 11